+++ src/render/program/program_uniforms.ts
@@ -11,6 +11,7 @@ import {symbolIconUniforms, symbolSDFUniforms, symbolTextAndIconUniforms} from '
 import {backgroundUniforms, backgroundPatternUniforms} from './background_program';
 import {terrainUniforms, terrainDepthUniforms, terrainCoordsUniforms} from './terrain_program';
 import {projectionErrorMeasurementUniforms} from './projection_error_measurement_program';
+import {atmosphereUniforms} from './atmosphere_program';
 
 const emptyUniforms = (_: any, __: any): any => {};
 
@@ -44,4 +45,5 @@ export const programUniforms = {
     terrainDepth: terrainDepthUniforms,
     terrainCoords: terrainCoordsUniforms,
     projectionErrorMeasurement: projectionErrorMeasurementUniforms,
+    atmosphere: atmosphereUniforms,
 };