+++ src/render/program/projection_program.ts
@@ -1,6 +1,6 @@
-import {Uniform1f, Uniform4f, UniformLocations, UniformMatrix4f} from '../uniform_binding';
+import {Uniform1f, Uniform3f, Uniform4f, UniformLocations, UniformMatrix4f} from '../uniform_binding';
 import {Context} from '../../gl/context';
-import {mat4} from 'gl-matrix';
+import {mat4, vec3} from 'gl-matrix';
 
 export type ProjectionPreludeUniformsType = {
     'u_projection_matrix': UniformMatrix4f;
@@ -15,7 +15,7 @@ export const projectionUniforms = (context: Context, locations: UniformLocations
     'u_projection_tile_mercator_coords': new Uniform4f(context, locations.u_projection_tile_mercator_coords),
     'u_projection_clipping_plane': new Uniform4f(context, locations.u_projection_clipping_plane),
     'u_projection_transition': new Uniform1f(context, locations.u_projection_transition),
-    'u_projection_fallback_matrix': new UniformMatrix4f(context, locations.u_projection_fallback_matrix)
+    'u_projection_fallback_matrix': new UniformMatrix4f(context, locations.u_projection_fallback_matrix),
 });
 
 export type ProjectionData = {